requirements/default.txt
fastapi==0.110.0
Jinja2==3.1.3
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-experimental==0.0.53
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langgraph==0.0.26
langsmith==0.1.13
openai==1.13.3
pydantic==2.6.3
python-decouple==3.8
uvicorn==0.27.1
